Second change batch.
Further revisions made to Model layer in order for hibernate to properly fit the previous database structure. Changed naming conventions in Service and DAO layers. Additionally changed types for DAO custom methods. FURTHER ACTION; create a updateble/insertable=false column/field in models to hold ID values from Objects. This way the previous working DAO methods can continue to be used as opposed to the current new methods requiring an entire object (more than the Id may be required in this case for there to be a matching object found).